backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
pydantic-settings
networkx
requests
